testing double data
