+++ components/themeWrapper.tsx
@@ -28,20 +28,6 @@ export default function ({ hideNavbar, children }: Props) {
         }
     })
 
-    useEffect(() => {
-        function prepareUrl(params) {
-            const url = new URL(location.href)
-            const queryParams = new URLSearchParams(location.search)
-            let customUrl = url.protocol + "//" + url.hostname + url.pathname.replace(/\/$/, '')
-            for (const paramName of params) {
-                const paramValue = queryParams.get(paramName)
-                if (paramValue) customUrl = customUrl + '/' + paramValue
-            }
-            return customUrl
-        }
-        plausible('pageview', { u: prepareUrl(['destNetwork', 'destAddress', 'lockNetwork']) })
-    }, [])
-
     const { addressSource } = useQueryState()
     return <div className={` ${addressSource} styled-scroll`}>
         <div className="invisible imxMarketplace"></div>
