+++ pages/_document.tsx
@@ -27,6 +27,11 @@ export default class extends Document {
           {process.env.NEXT_PUBLIC_VERCEL_ENV && <script
             defer
             data-domain={process.env.NEXT_PUBLIC_VERCEL_ENV == 'production' ? 'layerswap.io' : "testnet.layerswap.io"}
+            src="https://plausible.io/js/script.tagged-events.js"
+          />}
+          {process.env.NEXT_PUBLIC_VERCEL_ENV && <script
+            defer
+            data-domain={process.env.NEXT_PUBLIC_VERCEL_ENV == 'production' ? 'layerswap.io' : "testnet.layerswap.io"}
             src="https://plausible.io/js/script.manual.js"
           />}
           <script
